+++ Bloch/Artifacts/UnlockedPotentialArtifact.cs
@@ -35,17 +35,17 @@ internal sealed class UnlockedPotentialArtifact : Artifact, IRegisterable
 	}
 
 	public override List<Tooltip> GetExtraTooltips()
-		=> ModEntry.Instance.KokoroApi.Spontaneous.MakeAction(new ADummyAction()).AsCardAction.GetTooltips(DB.fakeState);
+		=> ModEntry.Instance.KokoroApi.Impulsive.MakeAction(new ADummyAction()).AsCardAction.GetTooltips(DB.fakeState);
 
 	private static void SpontaneousManager_TriggerAction_Begin_Postfix(CardAction __instance, State s, Combat c)
 	{
 		if (s.EnumerateAllArtifacts().FirstOrDefault(a => a is UnlockedPotentialArtifact) is not { } artifact)
 			return;
-		if (ModEntry.Instance.KokoroApi.Spontaneous.AsAction(__instance) is not { } spontaneousAction)
+		if (ModEntry.Instance.KokoroApi.Impulsive.AsAction(__instance) is not { } impulsiveAction)
 			return;
 
-		if (string.IsNullOrEmpty(spontaneousAction.Action.artifactPulse))
-			spontaneousAction.Action.artifactPulse = artifact.Key();
-		c.QueueImmediate(spontaneousAction.Action);
+		if (string.IsNullOrEmpty(impulsiveAction.Action.artifactPulse))
+			impulsiveAction.Action.artifactPulse = artifact.Key();
+		c.QueueImmediate(impulsiveAction.Action);
 	}
 }